Unified math functions usage.

+++ src/tools.cpp
@@ -31,10 +31,10 @@ VectorXd Tools::CalculateRMSE(const vector<VectorXd> &estimations,
 	//accumulate squared residuals
 	for(int i=0; i < estimations.size(); ++i) {
         VectorXd res = estimations[i] - ground_truth[i];
-        rmse(0) += pow(res(0),2);
-        rmse(1) += pow(res(1),2);
-        rmse(2) += pow(res(2),2);
-        rmse(3) += pow(res(3),2);
+        rmse(0) += powf(res(0),2);
+        rmse(1) += powf(res(1),2);
+        rmse(2) += powf(res(2),2);
+        rmse(3) += powf(res(3),2);
 	}
 
 	//calculate the mean
@@ -43,10 +43,10 @@ VectorXd Tools::CalculateRMSE(const vector<VectorXd> &estimations,
 
 	//calculate the squared root
 	// ... your code here
-	rmse(0) = sqrt(rmse(0));
-	rmse(1) = sqrt(rmse(1));
-	rmse(2) = sqrt(rmse(2));
-	rmse(3) = sqrt(rmse(3));
+	rmse(0) = sqrtf(rmse(0));
+	rmse(1) = sqrtf(rmse(1));
+	rmse(2) = sqrtf(rmse(2));
+	rmse(3) = sqrtf(rmse(3));
 
   cout << "RMSE: " << rmse << endl;
 	//return the result
@@ -65,8 +65,8 @@ MatrixXd Tools::CalculateJacobian(const VectorXd& x_state) {
 	float vx = x_state(2);
 	float vy = x_state(3);
 
-	float normSqrt = pow(px, 2) + pow(py, 2);
-	float norm = sqrt(normSqrt);
+	float normSqrt = powf(px, 2) + pow(py, 2);
+	float norm = sqrtf(normSqrt);
 
 	//check division by zero
 	if (normSqrt == 0) {
@@ -82,8 +82,8 @@ MatrixXd Tools::CalculateJacobian(const VectorXd& x_state) {
 	Hj(1,1) = px / normSqrt;
 	Hj(1,2) = 0;
 	Hj(1,3) = 0;
-	Hj(2,0) = py * (vx*py - vy*px) / pow(norm, 3);
-	Hj(2,1) = px * (vy*px - vx*py) / pow(norm, 3);
+	Hj(2,0) = py * (vx*py - vy*px) / powf(norm, 3);
+	Hj(2,1) = px * (vy*px - vx*py) / powf(norm, 3);
 	Hj(2,2) = px / norm;
 	Hj(2,3) = py / norm;
